requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
yt-dlp==2023.12.30
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.1